clean up code; rename components and files
