+++ service-catalog-ui/src/components/ServiceClassDetails/ServiceClassDescription/ServiceClassDescription.js
