+++ service-catalog-ui/src/components/ServiceClassDetails/ServiceClassInstancesTable/ServiceClassInstancesTable.js
